refactor : 페이지 구분
