+++ src/layouts/App.tsx
@@ -8,24 +8,24 @@ import { login } from '../redux/features/userSlice';
 import { positionAction } from '../redux/features/postionSlice';
 import LoadingSpinner from '../components/loadingSpinner';
 
-const Home = React.lazy(() => import('../pages/Home'));
-const Splash = React.lazy(() => import('../pages/Splash'));
-const Login = React.lazy(() => import('../pages/Login'));
-const SignUp = React.lazy(() => import('../pages/SignUp'));
-const RedirectKakao = React.lazy(() => import('../pages/RedirectKakao'));
-const FailKakao = React.lazy(() => import('../pages/FailKakao'));
-const ChatListPage = React.lazy(() => import('../pages/ChatListPage'));
-const Search = React.lazy(() => import('../pages/Search'));
-const ChatDetail = React.lazy(() => import('../pages/ChatDetail'));
-const Profile = React.lazy(() => import('../pages/Profile'));
-const UserMatchMore = React.lazy(() => import('../pages/UserMatchMore'));
-const ProfileEdit = React.lazy(() => import('../pages/ProfileEdit'));
-const Maps = React.lazy(() => import('../pages/Maps'));
-const Match = React.lazy(() => import('../pages/Match'));
-const New = React.lazy(() => import('../pages/New'));
-const Keyword = React.lazy(() => import('../pages/keyword'));
-const MatchHistory = React.lazy(() => import('../pages/MatchHistory'));
-const Searching = React.lazy(() => import('../pages/Searching'));
+const Home = React.lazy(() => import('../pages/Home/Home'));
+const Splash = React.lazy(() => import('../pages/Auth/Splash'));
+const Login = React.lazy(() => import('../pages/Auth/Login'));
+const SignUp = React.lazy(() => import('../pages/Auth/SignUp'));
+const RedirectKakao = React.lazy(() => import('../pages/Kakao/RedirectKakao'));
+const FailKakao = React.lazy(() => import('../pages/Kakao/FailKakao'));
+const ChatListPage = React.lazy(() => import('../pages/Chat/ChatListPage'));
+const Search = React.lazy(() => import('../pages/Match/Search'));
+const ChatDetail = React.lazy(() => import('../pages/Chat/ChatDetail'));
+const Profile = React.lazy(() => import('../pages/Profile/Profile'));
+const UserMatchMore = React.lazy(() => import('../pages/Profile/UserMatchMore'));
+const ProfileEdit = React.lazy(() => import('../pages/Profile/ProfileEdit'));
+const Maps = React.lazy(() => import('../pages/Maps/Maps'));
+const Match = React.lazy(() => import('../pages/Match/Match'));
+const New = React.lazy(() => import('../pages/New/New'));
+const Keyword = React.lazy(() => import('../pages/Match/keyword'));
+const MatchHistory = React.lazy(() => import('../pages/Profile/MatchHistory'));
+const Searching = React.lazy(() => import('../pages/Match/Searching'));
 
 const App = () => {
   const user = useSelector((state: RootState) => state.user);
