+++ src/pages/Auth/Splash.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import { Link } from 'react-router-dom';
-import { SERVER_URL } from '../apis';
+import { SERVER_URL } from '../../apis';
 
 const REST_API_KEY = process.env.REACT_APP_REST_API_KEY;
 const REDIRECT_URI = process.env.REACT_APP_REDIRECT_URI;
