+++ src/pages/Chat/ChatDetail.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import Chat from '../components/Chat';
-import ChatTimeLine from '../components/ChatTimeLine';
-import ChatForm from '../components/ChatForm';
-import { useSocket } from '../hooks/useSocket';
+import Chat from '../../components/Chat';
+import ChatTimeLine from '../../components/ChatTimeLine';
+import ChatForm from '../../components/ChatForm';
+import { useSocket } from '../../hooks/useSocket';
 import { useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { RootState } from '../redux/store';
+import { RootState } from '../../redux/store';
 import { useQuery } from '@tanstack/react-query';
-import { apis } from '../apis';
+import { apis } from '../../apis';
 import dayjs from 'dayjs';
 import { v4 as uuidv4 } from 'uuid';
 import { Helmet } from 'react-helmet';
