+++ src/pages/Match/Match.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
-import GetPostList from '../components/GetPostList';
-import GuestPostList from '../components/GuestPostList';
+import GetPostList from '../../components/GetPostList';
+import GuestPostList from '../../components/GuestPostList';
 
 const Match: FC = () => {
   if (localStorage.getItem('accessToken') === null || '' || undefined) {
