+++ src/pages/New/New.tsx
@@ -1,18 +1,18 @@
 import React, { FC, useEffect, useState, useCallback, useRef } from 'react';
 import { useForm } from 'react-hook-form';
 import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../redux/store';
+import { RootState } from '../../redux/store';
 import { useNavigate, useLocation } from 'react-router-dom';
-import { apis } from '../apis';
-import { PostEditDataProps, ImageType } from '../typings';
-import MapContainer from '../components/MapContainer';
-import Calendars from '../components/Calendar';
-import CustomSubject from '../components/CustomSelect';
+import { apis } from '../../apis';
+import { PostEditDataProps, ImageType } from '../../typings';
+import MapContainer from '../../components/MapContainer';
+import Calendars from '../../components/Calendar';
+import CustomSubject from '../../components/CustomSelect';
 import { IoMdCloseCircleOutline } from 'react-icons/io';
-import { toggleCalendarShow, toggleClear, toggleModalShow, toggleSubjectShow } from '../redux/features/toggleSlice';
-import { addressClear } from '../redux/features/addressSlice';
-import { calendarClear } from '../redux/features/calendarSlice';
-import { subjectClear } from '../redux/features/subjectSlice';
+import { toggleCalendarShow, toggleClear, toggleModalShow, toggleSubjectShow } from '../../redux/features/toggleSlice';
+import { addressClear } from '../../redux/features/addressSlice';
+import { calendarClear } from '../../redux/features/calendarSlice';
+import { subjectClear } from '../../redux/features/subjectSlice';
 import { useQueryClient } from '@tanstack/react-query';
 
 const Newpost: FC = () => {
