+++ src/pages/Profile/ProfileEdit.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { RootState } from '../redux/store';
+import { RootState } from '../../redux/store';
 import { Helmet } from 'react-helmet';
-import UserNicknameChange from '../components/UserNIcknameChange';
+import UserNicknameChange from '../../components/UserNIcknameChange';
 
 interface UpdateUser {
   nickname: string;
