+++ src/util/constant/index.ts
@@ -1,0 +1,1 @@
+export const MY_CONSTANT = 'my constant';
